requirements.txt
requests
pandas
pyjwt
schedule
numpy
